backend/requirements.txt
nodriver
selenium_driverless
openai
browserforge
camoufox
python-dotenv
pythonbible
supabase
pandas
camoufox[geoip] 
fastapi[standard]
ruff
black
psycopg2-binary
